Move GetPodPriority from /scheduler/util to /api/pod
