+++ pkg/scheduler/util/BUILD
@@ -15,7 +15,7 @@ go_test(
     ],
     embed = [":go_default_library"],
     deps = [
-        "//pkg/apis/scheduling:go_default_library",
+        "//pkg/api/v1/pod:go_default_library",
         "//staging/src/k8s.io/api/core/v1:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/util/diff:go_default_library",
     ],
@@ -31,7 +31,7 @@ go_library(
     ],
     importpath = "k8s.io/kubernetes/pkg/scheduler/util",
     deps = [
-        "//pkg/apis/scheduling:go_default_library",
+        "//pkg/api/v1/pod:go_default_library",
         "//pkg/scheduler/api:go_default_library",
         "//pkg/scheduler/metrics:go_default_library",
         "//staging/src/k8s.io/api/core/v1:go_default_library",
